@@ -12,32 +12,39 @@
     "fmt": "prettier --write 'src/**/*.{ts,vue,json,scss}'",
     "fmt:check": "prettier --check 'src/**/*.{ts,vue,json,scss}'",
     "typecheck": "tsc --noEmit",
+    "test": "vitest",
+    "test:ui": "vitest --ui",
+    "test:coverage": "vitest --coverage",
     "prepare": "husky"
   },
   "dependencies": {
     "@nativescript/core": "~8.9.0",
+    "@nativescript/secure-storage": "^4.0.1",
     "@nativescript/theme": "^3.1.0",
     "nativescript-audio": "~6.2.0",
-    "nativescript-vue": "~2.9.3",
+    "nativescript-vue": "^3.0.1",
     "ofetch": "^1.3.4",
-    "vuex": "^3.6.2"
+    "vue": "^3.5.22",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@nativescript/types": "~8.9.0",
     "@nativescript/webpack": "~5.0.0",
-    "@types/node": "~17.0.21",
+    "@types/node": "^22.18.10",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
+    "@vitejs/plugin-vue": "^6.0.1",
+    "@vitest/ui": "^3.2.4",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^9.1.2",
     "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-vue": "^9.33.0",
+    "happy-dom": "^20.0.2",
     "husky": "^9.1.7",
     "lint-staged": "^15.5.2",
-    "nativescript-vue-template-compiler": "~2.9.3",
     "prettier": "^3.6.2",
     "typescript": "~5.4.0",
-    "vue": "~2.6.12"
+    "vitest": "^3.2.4"
   },
   "lint-staged": {
     "*.{ts,vue}": [
