@@ -1,13 +1,24 @@
-import Vue from 'nativescript-vue'
+/**
+ * NativeScript-Vue 3.x 应用入口
+ */
+
+import { createApp } from 'nativescript-vue'
 import InputView from './views/InputView.vue'
 import store from './stores'
 
 declare let __DEV__: boolean
 
-// Prints Vue logs when --env.production is *NOT* set while building
-Vue.config.silent = !__DEV__
+// 创建Vue 3应用实例
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const app = createApp(InputView as any)
 
-new Vue({
-  store,
-  render: (h: any) => h('frame', [h(InputView)]),
-}).$start()
+// 注册Vuex store
+app.use(store)
+
+// 开发模式配置
+if (__DEV__) {
+  app.config.performance = true
+}
+
+// 启动应用
+app.start()
