@@ -1,18 +1,13 @@
 /**
- * Vuex Store 类型定义
+ * Vuex 4 + Vue 3 类型定义
  */
 
 import { Store } from 'vuex'
 import { RootState } from '@/stores'
 
-declare module 'vue/types/vue' {
-  interface Vue {
+// 扩展 ComponentCustomProperties 以支持 $store
+declare module '@vue/runtime-core' {
+  interface ComponentCustomProperties {
     $store: Store<RootState>
-  }
-}
-
-declare module 'vue/types/options' {
-  interface ComponentOptions<_V extends Vue> {
-    store?: Store<RootState>
   }
 }
